chore: security code style 정리 및 move security const 파일
