+++ src/main/kotlin/com/example/attendanceapimono/adapter/infra/security/SecurityConfig.kt
@@ -36,8 +36,6 @@ class SecurityConfig(
     fun securityWebFilterChain(http: ServerHttpSecurity): SecurityWebFilterChain {
 
         return http
-//            .exceptionHandling()
-
             .csrf().disable()
             .formLogin().disable()
             .httpBasic().disable()
