+++ src/main/kotlin/com/example/attendanceapimono/adapter/infra/security/TokenProvider.kt
@@ -1,6 +1,5 @@
 package com.example.attendanceapimono.adapter.infra.security
 
-import com.example.attendanceapimono.constant.SecurityConst
 import com.example.attendanceapimono.util.log
 import com.example.attendanceapimono.util.tryOrNull
 import io.jsonwebtoken.*;
@@ -8,7 +7,6 @@ import io.jsonwebtoken.security.Keys
 import io.jsonwebtoken.security.SignatureException
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.Authentication
-import org.springframework.security.core.GrantedAuthority
 import org.springframework.security.core.authority.SimpleGrantedAuthority
 import org.springframework.security.core.userdetails.User
 import org.springframework.security.core.userdetails.UserDetails
@@ -17,12 +15,9 @@ import java.security.Key
 import java.util.*
 
 @Component
-class TokenProvider(val authConfig: AuthProperties) {
-    val key: Key;
-
-    init {
-        this.key = Keys.hmacShaKeyFor(this.authConfig.secret.toByteArray());
-    }
+class TokenProvider(authConfig: AuthProperties) {
+    private val key: Key = Keys.hmacShaKeyFor(authConfig.secret.toByteArray())
+    private val expires = authConfig.expires
 
     private fun getClaims(token: String) = Jwts
         .parserBuilder()
@@ -32,8 +27,8 @@ class TokenProvider(val authConfig: AuthProperties) {
         .body;
 
     fun createToken(userDetails: UserDetails): String {
-        val now: Date = Date();
-        val expiredDate: Date = Date(now.time + this.authConfig.expires)
+        val now = Date()
+        val expiredDate = Date(now.time + expires)
 
         return Jwts.builder()
             .setSubject(userDetails.username)
@@ -45,18 +40,14 @@ class TokenProvider(val authConfig: AuthProperties) {
             .compact();
     }
 
-    fun getUserId(token: String): UUID = UUID.fromString(this.getClaims(token).subject);
+    fun getUserId(token: String): UUID = UUID.fromString(this.getClaims(token).subject)
 
-    fun getAuthentication(token: String): Authentication? = tryOrNull {
+    fun getAuthentication(token: String) = tryOrNull<Authentication> {
         val claims = Jwts.parserBuilder()
             .setSigningKey(key)
             .build()
             .parseClaimsJws(token)
             .body
-
-//        if (claims[CLAIMS_TOKEN_TYPE] != ACCESS_TOKEN_TYPE) {
-//            return null
-//        }
 
         val authorities = (claims[SecurityConst.CLAIMS_ROLE] as ArrayList<*>).map { SimpleGrantedAuthority(it as String) }
 
@@ -69,18 +60,18 @@ class TokenProvider(val authConfig: AuthProperties) {
 
     fun validateToken(authToken: String): Boolean {
         try {
-            Jwts.parserBuilder().setSigningKey(this.key).build().parseClaimsJws(authToken);
+            Jwts.parserBuilder().setSigningKey(this.key).build().parseClaimsJws(authToken)
             return true;
         } catch (ex: SignatureException) {
-            log.error("Invalid JWT signature");
+            log.error("Invalid JWT signature")
         } catch (ex: MalformedJwtException) {
-            log.error("Invalid JWT token");
+            log.error("Invalid JWT token")
         } catch (ex: ExpiredJwtException) {
-            log.error("Expired JWT token");
+            log.error("Expired JWT token")
         } catch (ex: UnsupportedJwtException) {
-            log.error("Unsupported JWT token");
+            log.error("Unsupported JWT token")
         } catch (ex: IllegalArgumentException) {
-            log.error("JWT claims string is empty.");
+            log.error("JWT claims string is empty.")
         }
         return false;
     }
